Fix: Ensure float32 dtype for TensorFlow operations
TensorFlow Conv2D requires float32 tensors but numpy operations
were upcasting to float64, causing errors across all platforms.

Fixed two sources of float64 upcasting:
1. Kernel normalization: Added .astype(np.float32) after reshape
2. Bias generation: Wrapped np.random.uniform() with np.float32()

This ensures all tensors passed to TF operations are float32.

+++ aeon/transformations/collection/convolution_based/rocketGPU/_rocket_gpu.py
@@ -226,12 +226,13 @@ class ROCKETGPU(BaseROCKETGPU):
                     )
 
                 # Reshape to TensorFlow format: (kernel_size, num_selected_channels, 1)
+                # Ensure float32 dtype is maintained after normalization
                 _convolution_kernel = _weights_reshaped.T.reshape(
                     _kernel_size, _num_channels, 1
-                )
+                ).astype(np.float32)
 
                 # 4. Bias (matches CPU line 193)
-                _bias = np.random.uniform(-1.0, 1.0)
+                _bias = np.float32(np.random.uniform(-1.0, 1.0))
 
                 # 5. Dilation (matches CPU lines 195-198)
                 if self.use_dilation:
